baitap2

+++ baitapso2/src/Main.java
@@ -4,10 +4,10 @@ public class Main {
     public static void main(String[] args) {
         Scanner CV = new Scanner(System.in);
         System.out.printf("Nhập chiều dài = ");
-        int CD = CV.nextInt();
+        double CD = CV.nextDouble();
 
         System.out.printf("Nhập chiều rộng = ");
-        int CR = CV.nextInt();
+        double CR = CV.nextDouble();
 
         System.out.println("Chu vi diện tích hình chữ nhật = " + (CD+CR)*2);
         System.out.println("Diện tích hình chữ nhật = " + CD*CR);
